+++ idp-server/src/test/java/de/gematik/idp/server/PairingControllerAccessTest.java
@@ -193,7 +193,9 @@ class PairingControllerAccessTest {
 
   private String buildAccessTokenString() {
     return "Bearer "
-        + accessToken.encrypt(idpEnc.getIdentity().getCertificate().getPublicKey()).getRawString();
+        + accessToken
+            .encryptAsNjwt(idpEnc.getIdentity().getCertificate().getPublicKey())
+            .getRawString();
   }
 
   @Test